+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.4] - Unreleased
+
+### Added
+
+- Endpoint resource:
+
+```terraform
+resource "neon_endpoint" "this" {
+  project_id = "bitter-meadow-966132"
+  branch_id  = "br-floral-mountain-251143"
+  type       = "read_write"
+}
+```
+
 ## [0.0.3] - 2023-01-07
 
 ### Added
@@ -25,7 +39,7 @@ resource "neon_project" "this" {
 }
 
 resource "neon_branch" "this" {
-  project_id = neon_project.foo.id
+  project_id = neon_project.this.id
   name       = "bar"
 }
 ```